list and server working now
